Fix trigger condition for the cherry-pick workflow
PR #8511 updated the cherry-pick workflow to use lable "Request for 20xx branch".
However, the workflow condition was not updated accordingly. The workflow
actions are still skipped when labels like "Request for 20xx branch" is set.

This change updated the trigger condition in the workflow yaml to match labels
start with "Request for 20".

Signed-off-by: Xin Wang <xiwang5@microsoft.com>