[dualtor][active-active] Fix `validate_active_active_dualtor_setup` (#8566)
Approach
What is the motivation for this PR?

In the dualtor-aa nightly run, some testcase failed at validate_active_active_dualtor_setup as the mux ports are not all active. The root cause is fixture remove_ip_addresses restarts the ptf ports, which will make the sockets open in icmp_responder in an error state.
Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
So let's do the folllowing two things:

1. in remove_ip_addresses fixture, let's restart icmp_responder after restarting ptf ports if icmp_responder is running.
2. in validate_active_active_dualtor_setup, let's add recover to restart both icmp_responder and nic_simulator if mux status check fails.

How did you verify/test it?

Signed-off-by: yenlu <yenlu@cisco.com>