Removed skip for password hardening test (#8581)
This PR removes the skip of password hardening test, as the issue #6428 is already fixed, and the test can run on 202205 and 202211 branches.

What is the motivation for this PR?
Remove the skip for password hardening test as the test can run on any branch (not only master)

How did you do it?
Removed the skip from test_mark_conditions.py

Signed-off-by: yenlu <yenlu@cisco.com>