Fix tunnel qos remap test (#8561)
Fix #6667
This PR is to fix regression in tunnel_qos_remap.py.

Below test cases are fixed:
* test_encap_dscp_rewrite
* test_bounced_back_traffic_in_expected_queue
* test_tunnel_decap_dscp_to_queue_mapping
* test_separated_qos_map_on_tor
* test_pfc_pause_extra_lossless_standby
* test_pfc_pause_extra_lossless_active
* test_tunnel_decap_dscp_to_pg_mapping

Test case test_xoff_for_pcbb is not fixed yet. Will fix it in another PR.
The change of PR #8424 is also included in this PR.

What is the motivation for this PR?
This PR is to fix test regression in tunnel_qos_remap.py.

How did you do it?
Improve test logic and code to make it compatible with the multi-ASIC change.

How did you verify/test it?
The change is verified on a dualtor testbed.