Handle MegaBits speed for show int status <interface> output in test_update_speed testcase (#7981)
What is the motivation for this PR?
Test case 'test_update_speed' in test_eth_interface.py fails, when the supported selected speed to apply on the interface is in 'Mbps' speed instead of 'Gbps'.

How did you do it?
If the current status speed after applying the patch is in 'Mbps', handle it as it is done for 'Gbps' speed.

How did you verify/test it?
Ran the 'test_update_speed' tests against a T2 chassis.

Co-authored-by: rawal <amol.rawal@nokia.com>