Update testcase test_nhop_group_member_count
The intention of the testcase is to fill the ecmp member table,
but the testcase considers available ecmp groups and then adds
members to it without considering the availability in the member table.
The change is to fill the members based on availability of both the
member table and group table to avoid either of them going out of resource.

Changes:
1. For nhop_group_count, take the min of (availble ecmp groups) and
   (available ecmp members / default_max_nhop_paths).
2. Creating ARP entries is consuming some ecmp members. So, query the
   crm available resources after arp entries are programmed and synced.