Add Secure Upgrade Test (#6816)
What is the motivation for this PR?
We want to add a new secure upgrade test to validate non successful upgrade to non secure image.
In details:
If we have a secured image installed on a secured system, trying to install a non-secure image on it
should fail and we should expect a relevant message indicating so.

How did you verify/test it?
By taking a secured system with secured image installed on it and a path to non secure image we created privately,
we ran the test.

Supported testbed topology if it's a new test case?
Any topology is supported for the test.

Documentation
link to feature HLD: sonic-net/SONiC#1024

Signed-off-by: yenlu <yenlu@cisco.com>