Cisco-8000: multi-asic: qos-sai: Modifications to BufferPoolWatermark for cisco-8000 multi-asic. (#8541)
For cisco-8000 the fabric card between the source and dest asic brings in uncertainties regarding the number of packets in flight. So we focus on dest asic for BufferPoolWatermark test.

Modifications include:
Read the watermark only from dest asic during the test for cisco-8000.
At the end of the test, when we want to overflow the shared-pool, we use the watermark before the tx is enabled - because once the tx is enabled on dst port, the packets that were held up in the fabric make it through and the watermark jumps way above the limit we expect.