Leverage the fanout's ASIC to send PFC storm
Currently, the PFC storm is sent using packet socket on the host CPU, which causes some drawbacks

- There is strict timing requirement, which can not be guaranteed
- The more ports under PFC storm, the higher probability that the timing requirement is not satisfied

We will leverage the fanout's ASIC to send PFC storm

- Use SDK API to enable lossless traffic, set xoff to 0, and setup mappings for the PGs and the corresponding IEEE priorities

Limitation

Still keep sending pfc frame with cpu in  the storm docker, but rename it to pfc_gen_cpu.py in the docker.

If use the ASIC to send the pfc frame, ASIC could not send specified count of the pfc frame. need to still use the CPU to send the pfc frame.such as the test_pfc_counter test

Signed-off-by: Stephen Sun <stephens@nvidia.com>