cisco-8000: qos-sai: Skip LossyQueueVoq and pgShared watermark for  t2 for time-being. (#8480)
There are several testcases that are not applicable, or not work as they are written, when run on multiple-ASIC(or multi-DUT). We need to skip them for cisco-8000. These are:

LossyQueueVoq
voq1 ==> split voq (not applicable for T2) or fair voq (applicable for T2 downlink LC)
voq2 ==> default voq so we need to add on uplink LC (need to add back)
So we need to add it back as enhancement.

PG Shared Watermark Lossy for multi-asic and multi-dut scenarion
Reason for skip: In packet chassis lossy traffic will get dropped at egress ASIC so we cannot validate PG Watermark on ingress ASIC