Fix span test issue
In the span test, it will choose some physical port from the vlan, when there are more than one vlan, in original logic, it will always choose port from the first vlan member, but we can not guarantee there are more than 3 physical port in the first vlan member. change the logic to choose the physical port from the vlan which has more than 3 physical ports.

Change-Id: I587d716b7381b3cf9d44b09acc0120f078b260ee