[202012]Restore vlan running config after module test_host_vlan.py running. (#8763)
What is the motivation for this PR?
During module vlan/test_host_vlan.py running, it modify the vlan mac in running config. In teardown perid, it reset the vlan mac and add the vlan mac in running config, which is not exist in previous running config. We delete the vlan mac running config in the teardown period in this PR, which will avoid config reload and save running time.

How did you do it?
Delete the vlan mac running config in teardown period.

How did you verify/test it?
02:03:12 conftest.core_dump_and_config_check      L2085 INFO   | Core dump and config check passed for vlan/test_host_vlan.py