Skip dualtor ecn test cases on Nvidia platforms.
In the dualtor scenario, when dscp remapping is enabled and packet with dscp 2 or 6 is received by the standby tor, the mapping behavior is different between Nvidia and other vendors, so the test cases fail on Nvidia platforms.
Actually, this is not considered as a valid use case, the dscp 2 and 6 should be reserved for the remapped packets. And this has been confirmed with MSFT, so skip the cases for dscp 2 and 6 on Nvidia platforms when dscp remapping is enabled.

Change-Id: Icdf19a7191d25c36bdd4cdc659e96297f924abf7