sonic-mgmt: Fix test_memory_checker.py test for low memory docker in ram systems
The test_memory_checker.py test tries to consume >400MB of RAM in the
telemetry container using 'stress -m 6' which by default mallocs 6*256MB.

With docker in ram on a system with 4GB of RAM there is limited available memory
and we can no longer malloc 6*256MB.

Instead, modify the test to use 2 vm_workers which results in 'stress -m 2'
which gets us over the 400MB limit that is being checked by monit and the
config is in /etc/monit/conf.d/monit_telemetry. The test also does not wait
long enough before checking that the expected log messages are available
in syslog. Increase the time before checking syslog from 130sec to 360sec.
In testing it can take ~3min30sec for the logs to appear.

Signed-off-by: Aaron Payment <aaronp@arista.com>