Merged PR 3055: Skip service warm restart test on 202205 branch
Skip service warm restart test on 202205 branch.

The PR is to skip `test_service_warm_restart` on `202205` branch.
The test set the warm_restart flag of container to true, but didn't clear it at the end. That left the DUT at unhealthy state.
Althouth PR  https://github.com/sonic-net/sonic-mgmt/pull/8781 had added the cleanup operation to reset the warm_restart flag, we still want to skip the test as it's not required internally,