[chassis] update the T3 device type (#8779)
Summary:
Fixes #8700

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Back port request
 201911
 202012
[x ] 202205
Approach
What is the motivation for this PR?
The cable length is calculated based on the neighbor type and device type. In the T2 topology the uplink neighbor type in the minigraph is defined as CoreRouter. This causes the device to pick incorrect cable length for uplink neighbors.

Example
https://github.com/sonic-net/sonic-buildimage/blob/aa46167fdd09b5f973d6c39529715ad75a2a58ec/device/arista/x86_64-arista_7800r3a_36d2_lc/Arista-7800R3A-36D2-C36/0/buffers_defaults_t2.j2#L3-L9

How did you do it?
This PR fixes the neighbor type of uplink neighbors to be set to RegionalHub or AZNGHub to match the values defined in the buffer_default_t2.j2