Remove xfail condition mark for test_system_health (#8833)
What is the motivation for this PR?
system_health/test_system_health.py::test_service_checker_with_process_exit is running to pass with latest 202012 image whatever platform is.
But from test result, it's still xfail, because the build_version is 20201231.105, string comparing with 20201231.44, it's smaller than 20201231.44.

How did you do it?
Split the version number and do compare then.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>