added flag plt_reboot_ctrl_overwrite in reboot function to decide whether we want to overwrite wait and timeout.
added a flag plt_reboot_ctrl_overwrite in reboot function to decide whether we want to overwrite wait and timeout. by default is to overwrite, in this test test_reload_configuration_checks we don't want to wait for that long, we want to catch transient error, so the flag is False.