Remove key `WARM_RESTART` after module `bgp/test_bgp_slb.py` running. (#8784) (#8849)
Description of PR
There is a warm reboot in module bgp/test_bgp_slb.py, and it will add the key WARM_RESTART in running config to set teamsyncd_timer. This cause the inconsistent in running config before and after the test case running. In this pr, we delete the key WARM_RESTART in running after the module running, which will avoid unnecessary config reload and save running time.

What is the motivation for this PR?
There is a warm reboot in module bgp/test_bgp_slb.py, and it will add the key WARM_RESTART in running config to set teamsyncd_timer. This cause the inconsistent in running config before and after the test case running. In this pr, we delete the key WARM_RESTART in running after the module running, which will avoid unnecessary config reload and save running time.

How did you do it?
Use configlet to delete the key after module running.

How did you verify/test it?
03:22:45 conftest.core_dump_and_config_check      L2108 INFO   | Core dump and config check passed for bgp/test_bgp_slb.py

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>
Co-authored-by: jianquanye@microsoft.com