When dut is single-asic, set the key "asic0" to keep up with multi-asic. (#8884)
What is the motivation for this PR?
In PR (#6527), it enhanced function core_dump_and_config_check to be multi-asic aware. But in single-asic scenerio, it simply set the key None, which does not make scene. In this PR, I reset the key "asic0" in single-asic scenerio to keep consistent with the key value of multi-asic scenerio.

How did you do it?
Change the key in single-asic scenerio from None to asic0.

Signed-off-by: yenlu <yenlu@cisco.com>