Delete interfaces admin_state after module `vlan/test_autostate_disabled.py` running.  (#8901)
Description of PR
In module vlan/test_autostate_disabled.py, it will change the admin_state of interfaces. Although it will restore all interfaces to their original admin_state, it will add some entries in running config, which are not exist before. So in this PR, after the module running, we will delete these additional entries.

What is the motivation for this PR?
In module vlan/test_autostate_disabled.py, it will change the admin_state of interfaces. Although it will restore all interfaces to their original admin_state, it will add some entries in running config, which are not exist before. So in this PR, after the module running, we will delete these additional entries.

How did you do it?
Delete additional entries which are not exist before.

How did you verify/test it?
06:13:11 conftest.core_dump_and_config_check      L2131 INFO   | Core dump and config check passed for vlan/test_autostate_disabled.py

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>