Bug fix of arp_responder (#8740)
Summary:
This import is necessary in the arp_responder but was accidentally removed by PR #8064.
    import scapy.arch.pcapdnet
Need to add it back or some tests will fail.(we have observed failures in acl test which is related to this issue)
For details of this import, please reference this PR #8481

What is the motivation for this PR?
Fix the ansible arp_responder

How did you do it?
Add the import

import scapy.arch.pcapdnet
How did you verify/test it?
Run the affected acl test, all passed after the fix.

Co-authored-by: Chuan Wu <chuanw@nvidia.com>
Signed-off-by: yenlu <yenlu@cisco.com>