[202012]Remove multi-asic scenerio in fixture `core_dump_and_config_check` in 202012 branch.  (#8980)
Description of PR
In 202012 branch, there is no multi-asic scenerio. But in PR (#8886) and (#8880), they innovate multi-asic scenerio in fixture core_dump_and_config_check. We think this innovation is unnecessary, so in this PR, we delete the code related to multi-asic scenerio.

What is the motivation for this PR?
In 202012 branch, there is no multi-asic scenerio. But in PR (#8886) and (#8880), they innovate multi-asic scenerio in fixture core_dump_and_config_check. We think this innovation is unnecessary, so in this PR, we delete the code related to multi-asic scenerio.

How did you do it?
Remove the code related to multi-asic scenerio in fixture core_dump_and_config_check.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>