Enhance setup-management-network.sh to avoid break lab server connectivity
Lab server could have management IP configured on bridge "br1".
If run this script on such lab server, the original mgmt IP could
be changed and connectivity to the lab server may be lost.

This change updated the setup-management-network.sh script to
skip configuring mgmt IP for br1 bridge if it already exists.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>