Make cacl pass (#8994)
Description of PR
The expected rules for dualtor are present in the script as a fixture "ignore_hardcoded_cacl_rule_on_dualtor" but the global variable "ignored_iptable_rules" is not called anywhere. This is causing test_cacl_application_dualtor to fail because the script is not expecting these rules.