fib IP range length is function instead of attribute
The IP_decap_test.py PTF script needs to determine destination
IPs to cover based on fib IP range length. For the fib class, ip
range length is a function, not an attribute.

This change fixed the issue of calling `ip_range.length` to
`ip_range.length()`.

Without this fix, for ip ranges with length 1 or 2, they are
all considered as ip range with length longer than 2. Then
the first, last and a random IP within the range are tested.
Unnecessarily tested destination IPs repeatedly for such
IP ranges.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>