Fix deploying vsonic neighbors with multiple inventory files (#8943)
When deploying vsonic neighbors, the `config_db.json` that gets
generated for those neighbors now has a reference to the DUT's IP
address (in the DEVICE_NEIGHBOR_METADATA table). This is accessed
through `hostvars`, which is read from the inventory files passed in.

The problem is that if there are multiple inventory files being used,
then the add-topo and remove-topo commands will only pass in the
inventory file for the VMs, and not the DUTs. This results in `hostvars`
not containing the DUT's IP address, which results in a failure for the
`config_db.json` generation.

Fix this by also passing in the DUT's inventory file.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>