Enhance setup-management-network.sh to avoid break lab server connectivity (#8982)
What is the motivation for this PR?
Lab server could have management IP configured on bridge "br1". If run this script on such lab server, the original mgmt IP could be changed and connectivity to the lab server may be lost.

How did you do it?
This change updated the setup-management-network.sh script to skip configuring mgmt IP for br1 bridge if it already exists.

How did you verify/test it?
Tested on local server.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>