Cherry pick of #8089
What is the motivation for this PR?
To fix a rare issue when control plane is detected to be down before reboot is issued. This happens if any out of 10 ping packets are dropped (cause unknown). When this happens, the logic proceeds to wait (join) for sender and sniffer threads to complete. Since the device never really rebooted, there is a chance sender and sniffer threads are not initialized and started.

Additonally, removed the unused method send_and_sniff

How did you do it?
Fix the issue by:
Initializing sender and sniffer threads in init.
Before joining the IO threads wait for them to start -- keep checking isAlive()

How did you verify/test it?
Tested on physical testbed and fix works.