[202012][fib] Fix test_fib failed in Nokia 7215 (#8960) (#9078)
What is the motivation for this PR?
Manually cherry-pick and resolve conflict of this PR: #8960
test_fib will randomly fail in marvell asic. To fix it.

How did you do it?
It failed because "240.0.0.0/24" is not excluded in test ip ranges. This PR #7030 has added it but not let it take effect due to missing asic_type in the parameter.

How did you verify/test it?
Run tests.

Signed-off-by: Yaqiang Zhu <yaqiangzhu@microsoft.com>