[platform] Register platform/api tests options to root conftest (#9106)
What is the motivation for this PR?
To run nightly test with platform/api options, we need to register it to root conftest.

How did you do it?
Make platform_tests/api folder as a module
Register api module as a plugin in root conftest

How did you verify/test it?
Run tests.

Signed-off-by: Yaqiang Zhu <yaqiangzhu@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>