[GCU E2E Testing] Override is_valid_platform_and_version (#8945)
Due to the issue #8944, the RDMA Platform Validator PR was reverted. Because the validator is no longer in place, all platforms are deemed a valid platform and version. To account for this, this PR returns True for is_valid_platform_and_version so that we expect GCU to be applied successfully on all platforms.

After the issue #8944 is resolved and the Platform Validator PR is merged, this fix will be removed, and we will expect op success or failure accordingly.

What is the motivation for this PR?
The RDMA Platform Validator PR was reverted, so test case expectations need to be adjusted accordingly (all platforms are considered valid, so return True)