[advanced-reboot] Use nohup when executing reboot CLI to prevent PIPE errors (#9171)
What is the motivation for this PR?
In the device where logs are in tmpfs, a backup is made for log_analyzer to work after reboot.
However, there have been many cases when one or more backup files are empty or not created.
Based on current RCA this happens when reboot CLI is executed from test and then SSH command times out.

When command execution ends, but reboot is still in progress in the device, any writes to stdout / etderr fails as there is no reader.

How did you do it?
To fix no reader issue, run the reboot CLI w/ nohup and redirect the output to a file.

Tested on physical device