[master] Change tool scripts to use Python3 explicitly (#9184)
Summary:
This is a part of Python3 migration. The incoming new sonic-mgmt-docker will set Python3 as default.

What is the motivation for this PR?
This is a part of Python3 migration. The incoming new sonic-mgmt-docker will set Python3 as default.

How did you do it?
In tool scripts, add head: #!/usr/bin/env python2
In any caller of tool scripts, use "./" to call the scripts.

Signed-off-by: yenlu <yenlu@cisco.com>