cisco: qos-sai: Update all-port-q-wm to use the dscp->Q mapping from the DUT itself. (#8933)
The test testQosSaiQWatermarkAllPorts has a hardcoded dscp->queue mapping data. This PR updates it to read the mapping data from the DUT and use that for testing. Changes:

Update the test_qos_sai.py::testQosSaiQWatermarkAllPorts to read the dscp->queue mapping from DUT.
Pass the new parameter to the PTF script.
Change the PTF script to use the CLI argument for the mapping instead of the hard-coded value.