Skip test_module for non T2 topology (#9209)
Description of PR
Summary:
Fixes # (issue)
test_module related cases would get the value of “get_num_modules”, if the value is equal to zero, then the case would be skipped due to “No modules found on device”.
The value of “get_num_modules” on T0/T1 is zero, so the cases would be skipped.
But if there is something wrong with API or unhealthy testbed issue, get_num_modules will not return 0 for T0/T1.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>