[renumber-topo] Add extra mgmt ip to ptf redeploy step (#9230)
What is the motivation for this PR?
If a testbed supports extra mgmt IP then the PTF gets the extra IP when add-topo is run. This change was done as part of #8432
However, if PTF redeploy is done, the extra IP added by add-topo is lost as PTF is recreated.
This issue is fixed in this PR by adding extra mgmt IP support in renumber_topo ansible playbook.

How did you do it?

How did you verify/test it?
Tested on physical testbed.