Fix import issue of tests/snappi_tests in python3 (#9232)
What is the motivation for this PR?
To support smooth python3 migration, we need to explicitly specify the python version to be used in each branch. For master branch, default python should be python3. Need to use python3 to run the collect-only check as well.
Because of the gap in collect-only check of PR testing, some import issues under tests/snappi_tests are not caught.

How did you do it?
* Use "python3 -m pytest" instead of directly run "pytest" command to do collect-only check.
* Fixed all the import issues of tests/snappi_tests in python3.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>