Fix potential KeyError of fixtures like vmhost
The fixtures creating testbed device objects depend on supplied inventory files.
When multiple inventory files are supplied to the pytest command, fixture 'enhance_inventory'
ensures that a list of inventory files are supplied to ansible's InventoryManager under the hood.

Although fixture `enhance_inventory` is set to "autouse". There is still chance that it is executed
after fixtures like vmhost, ptfhost, fanouthots, etc. In this case, multiple inventory files may be
supplied to InventoryManager as a string using "," to separate the inventory files. Then
InventoryManager may fail to find the correct inventory files and eventually result in "KeyError"
while trying to return some hosts defined in the inventory files.

This change fixed the issue by specifying "enhance_inventory" in the argument list of
fixtures vmhost, fanouthosts, ptfhost, etc. This can ensure that fixture "enhance_inventory"
is always executed before the device objects creating fixtures. The ansible InventoryManager
can always get a list of inventory files instead of a string with "," in the middle.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>