Delete config `suppress-fib-pending` after module `bgp/test_bgp_update_timer.py` running.  (#9231)
Description of PR
PR #7907 import a fixture has_suppress_feature to get bgp suppress-fib-pending on DUT, which may add a config suppress-fib-pending in running config. And this will cause inconsiscient between previous running config and current running config, and will cause unnecessary config reload after whole module running. In this PR, I delete this config in the teardown period of this module.

What is the motivation for this PR?
PR #7907 import a fixture has_suppress_feature to get bgp suppress-fib-pending on DUT, which may add a config suppress-fib-pending in running config. And this will cause inconsiscient between previous running config and current running config, and will cause unnecessary config reload after whole module running. In this PR, I delete this config in the teardown period of this module.

How did you do it?
I delete this config in the teardown period of this module.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>