lag_test: Add check to verify that an initial LACP packet was received
In the LacpTimingTest class, when getting the initial LACP packet, first
check to see that a packet was captured before trying to convert the
timestamp value of the tuple to a float. Otherwise, there'll be an error
about an invalid argument to float.

This doesn't necessarily prevent an error from being thrown, it just
makes it a more graceful/expected failure.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>