Add try-finally to recovery DUT in module `acl/test_stress_acl.py`.  (#9331)
Description of PR
Module acl/test_stress_acl.py will add several acl tules in config to test stress, and will delete the stess rules at the end of TC. But if there is something wrong happens, it will go to the end of TC, and will not delete these rules. This will cause unnecessary config reload. In this PR, we add try-finally to ensure the logic of recovering will be exectued.

What is the motivation for this PR?
Module acl/test_stress_acl.py will add several acl tules in config to test stress, and will delete the stess rules at the end of TC. But if there is something wrong happens, it will go to the end of TC, and will not delete these rules. This will cause unnecessary config reload. In this PR, we add try-finally to ensure the logic of recovering will be exectued.

How did you do it?
Add try-finally to ensure the logic of recovering will be exectued.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>