Fix EVERFLOW egress mirroring test in dualtor setup (#9413)
Summary:
This PR is to backport #9414 into 202205 branch.
The EVERFLOW Egress mirroring test was failing on Mellanox dualtor because the expected mirroring packet doesn't match the sent packet.

pytest_assert(inner_packet.pkt_match(mirror_packet_sent), "Mirror payload does not match received packet")
It's because for the downstream traffic on dualtor testbed, the src_mac should be the MAC address of Vlan1000. The value is different from the router MAC.

This PR addressed the issue by changing the src MAC address of the downstream traffic to VLAN MAC on dualtor testbed.