Workaound for ansible module executed twice issue of pytest-ansible
This is to workaround pytest-ansible plugin issue:
    Even no extra-inventory is specified, extra_inventory_manager will still be initialized.
    https://github.com/ansible-community/pytest-ansible/issues/135

As of today, pytest-ansible supports host manager and module dispatcher v29, v212, v213.
While initializing the pytest-ansible plugin, it tries to collect default ansible configurations and
command line options. These options are used for creating host manager. When no extra_inventory is specified,
the options passed to host manager will include extra_inventory=None. However, the host manager will still
try to create extra_inventory_manager with None as the inventory source. This will cause the module dispatcher
to run the module on hosts matching host pattern in both inventory and extra inventory. This would cause an
ansible module executed twice on the same host. In case we wish to use the shell module to run command like
"rm <some_file>" to delete a file. The second run would fail because the file has been deleted in the first run.
For more details, please refer to the Github issue mentioned above.

The fix is to overwrite the pytest-ansible plugin's initialize method and remove the extra_inventory option
if it is None. This will prevent the host manager from creating extra_inventory_manager with None as the
inventory source.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>