[fdb] Address failures on td2 asics for fdb mac move test (#9487)
Signed-off-by: Neetha John <nejo@microsoft.com>

What is the motivation for this PR?
On certain platforms with low disk space, failures are seen occassionally on running the test_fdb_mac_move case. This testcase runs in a loop and the number of times the loop is executed is dependent on the completeness_level setting. Lot of syslog messages are generated per loop and for completeness_level settings of 'confident' and above, it can lead to disk full case for certain platforms. This PR resets the completeness_level to basic for td2 platforms so as to reduce the number of iterations of test run.

How did you verify/test it?
Ran the testcase multiple times against t0-backend topology and no longer see failures

Signed-off-by: yenlu <yenlu@cisco.com>