[icmp_responder] Cache replies to reduce response time (#9531)
Approach
What is the motivation for this PR?
As the probes received from one port are the same basically(except those with TLVs), we could simply cache the replies instead of generating them in each reply cycle.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Use functools.lru_cache to cache the calls to icmp_reply.

How did you verify/test it?
With the lru_cache, the execution time of 1000 repeated calls to icmp_reply is reduced from 0.6184256076812744s to 0.0006690025329589844s.

Any platform specific information?

Signed-off-by: yenlu <yenlu@cisco.com>