Fix PR #9312 for KeyError and AttributeError (#9523)
Description of PR
PR #9312 will try to get DATAACL in running config in fixture recover_acl_rule. But if param enable_data_acl is set false when deploying minigraph, DATAACL will not appear in running config. Which may cause key error in fixture recover_acl_rule. In this PR, we fix this issue.

What is the motivation for this PR?
PR #9312 will try to get DATAACL in running config in fixture recover_acl_rule. But if param enable_data_acl is set false when deploying minigraph, DATAACL will not appear in running config. Which may cause key error in fixture recover_acl_rule. In this PR, we fix this issue.

How did you do it?
Run TC on testbed which enable_data_acl is false when deploying minigraph.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>