[advanced-reboot][dualtor] Update Loopback0 ping test: expect same count for sent and rcvd packets (#9599)
Summary: Currently dual tor warmboot fails w/ control plane not ready for warmboot. This is because the test expects 2 responses for every ping packet sent to Loopback. This expectation is not correct after PR #9089 is merged.

Update the testcase to now only expect 1 icmp response for every icmp packet sent to Loopback0.
Updated the test to remove the earlier expectation of removing the packets that were sent by icmp_responder.

This essentially reverts part of older PR: #6968

How did you verify/test it?
Tested on a physical testbed and this test issue is not seen.
The test is now able to see the responses of ping packets