[SNMP][IPv6]: Skip SNMP IPv6 related tests in 202211 branch. (#9630)
What is the motivation for this PR?
Issue sonic-net/sonic-buildimage#16165

Changes added in the PR sonic-net/sonic-buildimage#15487 to support SNMP over IPv6 does not handle if mgmt IP address is obtained via DHCP.

How did you do it?
Skip the sonic-mgmt SNMP IPv6 related tests in 202211 so that the sonic-net/sonic-buildimage#15487 can be reverted.

Signed-off-by: yenlu <yenlu@cisco.com>