Merged PR 3386: Skip test_dns_resolv_conf.py on a public image and internal sonic-mgmt
Skip test_dns_resolv_conf.py on a public image and internal sonic-mgmt

This is because there's inconsistent expectations on what DNS
nameservers may be present with this particular combo (whether a config
reload or reboot has happened since the DNS nameserver was programmed by
sonic-mgmt-int, or not, for instance). Because of this, until there's a
more programmatic method more widely used, skip this combination.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>