Add testbed health check scripts for testbed automatic monitor and recover (#9765)
What is the motivation for this PR?
SONiC Elastictest is onboarding to support testbed health automatic monitor and recover. It needs to do some health checks on testbeds.

How did you do it?
Add a new script testbed_health_check.py.

How did you verify/test it?
Verify locally.

Example:

 ./testbed_health_check.py -t vms2-4-t0-2700 -i ../ansible/str --tbfile ../ansible/testbed.yaml -o testbed_health_check_result_vms2-4-t0-2700 --log-level info

It will print the basic testbed check result to the console, if you specify an output file, it will write to it as well.
---------

Signed-off-by: Chun'ang Li <chunangli@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>