Add bgp sentinel support in GCU (#9735)
What is the motivation for this PR?
This PR is to verify the usage of 'config apply-patch' works on bgp sentinel config

How did you do it?
Add bgp sentinel config to dut and check if config change as expected

How did you verify/test it?
Run test of sonic-mgmt/tests/generic_config_updater/test_bgp_sentinel.py on KVM

Signed-off-by: yenlu <yenlu@cisco.com>